Latvian: two important fixes regarding V2 clauses
Agreement with the focus part (the object) works not only for Pers3, but also for Pers1 and Pers2 NPs (if the verb requires non-typical subject/object valences).
Object-dependent double negation works (in addition to the subject-dependent double negation).
The first one is motivated by PhrasebookLav, the second one - by AttemptoLav.
Also a couple of minor fixes.