Maltese: finished morphology for derived verbs
Of course it's never really finished. There's more cases I should write treebanks
for and test, but I think the coverage is good enough and the implementation
for the verbs is solid enough that only minor fixes may arise.

There's also the issue of verb participles, but I'll come back to those
when I am looked more closely at the syntax.