Wide Coveage Translation Demo: sort parse results by probability Because of the "heurisitc factor", the results returned by the parser might not be in strict probability order. To compensate, a sorting pass has been added to find and show the translation with the highest probability among the 10 first translations.
However, this means that the translation demo now immediately has to ask for
10 translations of every segment. Before, it initially asked for only one
translation of every segment, and then 9 more translations for a segment
when/if the user clicked on it. This change can slow down translation
noticeably (e.g. from 15s 30s to load an example with fiction in English).