Add a Travis configuration for autotesting commits
The configuration uses Nix to install dependencies.  I first tried using
just Cabal, but compiling all the dependencies takes too long for the
free tier of Travis.  Nix doesn't compile Hackage dependencies because
it has a "binary cache" with compiled versions of almost all
Hackage packages.

I do not run the GF test suite, because it seems to be in a
failing state.  This can be enabled later.

All of this will only be enabled once the Travis service is set up by
someone in the GF team with the permissions to do so.

Build results will then be automatically visible on all new commits and
pull requests.